+++ dpdk/examples/quota_watermark/qw/init.h
@@ -34,9 +34,9 @@
 #ifndef _INIT_H_
 #define _INIT_H_
 
-void configure_eth_port(uint8_t port_id);
+void configure_eth_port(uint16_t port_id);
 void init_dpdk(void);
-void init_ring(int lcore_id, uint8_t port_id);
+void init_ring(int lcore_id, uint16_t port_id);
 void pair_ports(void);
 void setup_shared_variables(void);
